refactor(icons/members): various fixes regarding icons and members
also changed the file structure of some directories
