+++ app/hooks/Icons/useGetMemberRoleIcons.tsx
@@ -9,7 +9,9 @@ export const useGetMemberRoleIcons = () => {
       [MemberRole.MODERATOR]: (
         <ShieldCheck className="h-4 w-4 ml-2 text-indigo-500" />
       ),
-      [MemberRole.ADMIN]: <ShieldAlert className="h-4 w-4 text-rose-500" />,
+      [MemberRole.ADMIN]: (
+        <ShieldAlert className="h-4 w-4 text-rose-500 mr-2" />
+      ),
     }
   }, [])
 }
